Removed old player + cleanup
