+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["gonk", "gonk-database", "player", "gonk-player"]
+members = ["gonk", "gonk-database", "gonk-player"]
 
 [profile.release]
 strip = true
