+++ gonk-player/Cargo.toml
@@ -1,24 +1,9 @@
 [package]
-name = "gonk-old-player"
-version = "0.1.0"
+name = "gonk-player"
+version = "0.2.0"
 edition = "2021"
-authors = ["Bay"]
-description = "Music playback library for gonk"
-repository = "https://github.com/zX3no/gonk"
-readme = "README.md"
-license = "CC0-1.0"
 
 [dependencies]
-symphonia = {version = "0.5.1", features = ["mp3", "isomp4", "alac", "aac"] }
-gonk-database = {version = "0.1.0", path = "../gonk-database"}
-
-[target.'cfg(unix)'.dependencies]
-gag = "1.0.0"
-alsa = "0.6"
-nix = "0.24.1"
-libc = "0.2.65"
-jack = "0.10.0"
-
-[target.'cfg(target_os = "windows")'.dependencies]
-winapi = { version = "0.3", features = ["audioclient", "combaseapi", "devpkey", "errhandlingapi", "handleapi", "ksmedia", "mmdeviceapi", "synchapi", "winbase"] }
-once_cell = "1.12"
+symphonia = {version = "0.5.1", default-features = false,  features = ["flac", "mp3", "ogg"]}
+winapi = { version = "0.3.9", features = ["mmdeviceapi", "combaseapi", "devpkey", "audioclient", "synchapi"] }
+gonk-database = {version = "0.2.0", path = "../gonk-database"}
