--- gonk-player/src/cpal/host/mod.rs
@@ -1,8 +1,0 @@
-#[cfg(any(target_os = "linux", target_os = "dragonfly", target_os = "freebsd"))]
-pub(crate) mod alsa;
-
-#[cfg(any(target_os = "linux", target_os = "dragonfly", target_os = "freebsd"))]
-pub(crate) mod jack;
-
-#[cfg(windows)]
-pub(crate) mod wasapi;
