--- gonk-player/src/cpal/host/wasapi/mod.rs
@@ -1,68 +1,0 @@
-extern crate winapi;
-
-pub use self::device::{
-    default_input_device, default_output_device, Device, Devices, SupportedInputConfigs,
-    SupportedOutputConfigs,
-};
-pub use self::stream::Stream;
-use self::winapi::um::winnt::HRESULT;
-use crate::cpal::traits::HostTrait;
-use crate::cpal::BackendSpecificError;
-use crate::cpal::DevicesError;
-use std::io::Error as IoError;
-
-mod com;
-mod device;
-mod stream;
-
-#[derive(Debug)]
-pub struct Host;
-
-impl Host {
-    pub fn new() -> Result<Self, crate::cpal::HostUnavailable> {
-        Ok(Host)
-    }
-}
-
-impl HostTrait for Host {
-    type Devices = Devices;
-    type Device = Device;
-
-    fn is_available() -> bool {
-        // Assume WASAPI is always available on Windows.
-        true
-    }
-
-    fn devices(&self) -> Result<Self::Devices, DevicesError> {
-        match Devices::new() {
-            Ok(devices) => Ok(devices),
-            Err(err) => Err(DevicesError::BackendSpecific { err }),
-        }
-    }
-
-    fn default_input_device(&self) -> Option<Self::Device> {
-        default_input_device()
-    }
-
-    fn default_output_device(&self) -> Option<Self::Device> {
-        default_output_device()
-    }
-}
-
-#[inline]
-fn check_result(result: HRESULT) -> Result<(), IoError> {
-    if result < 0 {
-        Err(IoError::from_raw_os_error(result))
-    } else {
-        Ok(())
-    }
-}
-
-fn check_result_backend_specific(result: HRESULT) -> Result<(), BackendSpecificError> {
-    match check_result(result) {
-        Ok(()) => Ok(()),
-        Err(err) => Err(BackendSpecificError {
-            description: format!("{}", err),
-        }),
-    }
-}
