--- player/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "gonk-player"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-symphonia = {version = "0.5.1", default-features = false,  features = ["flac", "mp3", "ogg"]}
-winapi = { version = "0.3.9", features = ["mmdeviceapi", "combaseapi", "devpkey", "audioclient", "synchapi"] }
-
-gonk-database = {version = "0.1.0", path = "../gonk-database"}
